my-taxi-pipeline/pipeline/assets/ingestion/requirements.txt
pandas
pyarrow
requests
numpy
python-dateutil